* #37
increment version to beta5

* #37
reworked protocol (untested)
added some first ideas (commented) wrt. multiple GAs

Code compiles fine.

* #37
added/reworked EEPROM address defines

* includes=KonnektingDevice.h added

* Create keywords.txt

* bug fix for PA 1.1.1 on NCN5120

* some general DPTs added

1.000 general bool
5.000 general byte
7.000 general unsigned integer
8.000 general integer
9.000 general float
12.000 general unsigned long
13.000 general long

* ESP8266 bug fixed

* FreeRAM for SAMD21G18A added

* Changed SAMD21G18A to ARDUINO_ARCH_SAMD to support all variants of SAMD21

* memmoryCommit() now works not only with ESP8266

* ProgButton and ProgLED are now optional

* initial support for STM32 and ESP32

* Added constants for DPTs
smaller fixes
up2date commit since weeeeks.

* updated

* added

* fixed merge issue

* added templates byte-read

* fixed compile issues

* updated DPT constants

* fixed issue #46

* Replaced ActionRingBuffer with own version
Added ArrayList implementation

* removed obsolete code

* cleanup

* cleanup

* refactored function names of KnxComObject to lower case

* cleanup + better documentation

* updated DPT stuff

* updated readme, to reflect major code changes to becom independant

* reworked code, added better comments

* added first doxygen tests
swtch to beta4a as intermediate version before beta5; required for roto actuator...

* fixed wrong include

* working on params with integer index

* updated doxygen config

* updated doxygen config, added more documentation in doxygen style

* updated doxygen

* updated doxygen

* updated doxygen

* updated doxygen

* fixed broken code

* ProgLED and ProgButton pins are now optional

New init function for own ProgLED and ProgButton implementation
New DemoSketch for new init-function
DemoSketches updatet to beta4a

* Initial support for ESP32 and STM32

Demo Sketches renamed
Some changes for ESP* and STM32
ESP32 has timing problems... need to be checked
ESP8266 won't compile
STM32 not tested yet

* initial travis ci file

just started testing travis ci ... :-)

* added onewire dependency for examples

* added DallasTemperature lib added for examples

* added example sketch lib

Added "Added "SparkFun HTU21D Humidity and Temperature Sensor Breakout"" dependency for example sketch

* skipping this example for travis ci

* updated build platforms

* removed

* updated dependency

* Update .travis.yml

* Update Konnekting_communication_test.ino

* Update kdevice_Temperature_DS18B20.h

* Update kdevice_Temp_RH.h

* Update kdevice_DemoSketch.h

* fix softserial for zero

* adapted to beta4a

* fixed softserial for zero

* excluded Arduino Zero for now

* skipping for zero

* added missing dtostrf -> workaround for now

* removed "void" as argument

* possible fix of isActive() issue with esp8266

* added travis ci badge

* rework init function

ESP8266 can now compile and send telegrams, but still has problems with millis() and micros() -> telegram reception doesn't work.
STM32 still not tested yet

* updated doxygen info for new init-function

* some fixes

revert demo sketches update

* STM32 fixes and DemoSketches update (#52)

* STM32: workaround for va_start&va_end

va_start & va_end are not defined in STM32 core 1.2.0

* DemoSketches for beta4a updated

Communication sketch rewritten.
It can be used now with: 328p, 32u4, SAMD21, STM32, ESP8266

* zero support added

* added flash storage lib

* removed esp8266 build as core has timing issues

* updated, formatted

* added define for versioning

+++ examples/DemoSensor_Temp_RH/kdevice_Temp_RH.h
@@ -51,4 +51,4 @@ byte KonnektingDevice::_paramSizeList[] = {
     /* Index 13 - rhMaxValue */ PARAM_UINT8,
     /* Index 14 - rhMaxLimit */ PARAM_INT16
 };
-const byte KonnektingDevice::_numberOfParams = sizeof (_paramSizeList); // do not change this code
+const int KonnektingDevice::_numberOfParams = sizeof (_paramSizeList); // do not change this code
